implement common header component and change all pages accordingly
