+++ packages/apps/graph-client/src/pages/account/transfers/[module]/[account].tsx
@@ -1,10 +1,8 @@
 import { useGetTransfersQuery } from '@/__generated__/sdk';
-import Loader from '@components/loader/loader';
-import { mainStyle } from '@components/main/styles.css';
-import { Text } from '@components/text';
+import Loader from '@/components/Common/loader/loader';
+import { mainStyle } from '@/components/Common/main/styles.css';
 import routes from '@constants/routes';
 import { Box, Button, Link, Notification, Table } from '@kadena/react-ui';
-import Head from 'next/head';
 import { useRouter } from 'next/router';
 import React from 'react';
 
@@ -22,19 +20,7 @@ const AccountTransfers: React.FC = () => {
 
   return (
     <div style={{ padding: '0 50px 30px 50px' }}>
-      <Head>
-        <title>Kadena Graph Client</title>
-        <link rel="icon" href="/favicon.png" />
-      </Head>
-
       <main className={mainStyle}>
-        <Text
-          as="h1"
-          css={{ display: 'block', color: '$mauve12', fontSize: 48, my: '$12' }}
-        >
-          Kadena Graph Client
-        </Text>
-
         <div>
           {loading && (
             <div style={{ display: 'flex', alignItems: 'center' }}>
