+++ packages/apps/graph-client/src/pages/block/overview/[hash].tsx
@@ -2,13 +2,12 @@ import {
   useGetBlockFromHashQuery,
   useGetMaximumConfirmationDepthQuery,
 } from '@/__generated__/sdk';
+import Loader from '@/components/Common/loader/loader';
+import { mainStyle } from '@/components/Common/main/styles.css';
 import { CompactTransactionsTable } from '@components/compact-transactions-table/compact-transactions-table';
-import Loader from '@components/loader/loader';
-import { mainStyle } from '@components/main/styles.css';
 import { Text } from '@components/text';
 import routes from '@constants/routes';
 import { Accordion, Box, Notification, Table } from '@kadena/react-ui';
-import Head from 'next/head';
 import { useRouter } from 'next/router';
 import React from 'react';
 
@@ -23,19 +22,7 @@ const Block: React.FC = () => {
 
   return (
     <div>
-      <Head>
-        <title>Kadena Graph Client</title>
-        <link rel="icon" href="/favicon.png" />
-      </Head>
-
       <main className={mainStyle}>
-        <Text
-          as="h1"
-          css={{ display: 'block', color: '$mauve12', fontSize: 48, my: '$12' }}
-        >
-          Kadena Graph Client
-        </Text>
-
         <div>
           {loading && (
             // Display a loading spinner next to the text without a gap
