+++ packages/apps/graph-client/src/pages/event/[key].tsx
@@ -1,10 +1,8 @@
 import { useGetEventByNameSubscription } from '@/__generated__/sdk';
-import Loader from '@components/loader/loader';
-import { mainStyle } from '@components/main/styles.css';
-import { Text } from '@components/text';
+import Loader from '@/components/Common/loader/loader';
+import { mainStyle } from '@/components/Common/main/styles.css';
 import routes from '@constants/routes';
 import { Notification, Table } from '@kadena/react-ui';
-import Head from 'next/head';
 import { useRouter } from 'next/router';
 import React from 'react';
 
@@ -21,19 +19,7 @@ const Event: React.FC = () => {
 
   return (
     <div>
-      <Head>
-        <title>Kadena Graph Client</title>
-        <link rel="icon" href="/favicon.png" />
-      </Head>
-
       <main className={mainStyle}>
-        <Text
-          as="h1"
-          css={{ display: 'block', color: '$mauve12', fontSize: 48, my: '$12' }}
-        >
-          Kadena Graph Client
-        </Text>
-
         <div>
           {eventLoading && (
             <div style={{ display: 'flex', alignItems: 'center' }}>
