+++ packages/apps/graph-client/src/pages/transactions/index.tsx
@@ -1,11 +1,9 @@
 import { Box, Notification } from '@kadena/react-ui';
 
 import { useGetTransactionsQuery } from '@/__generated__/sdk';
+import Loader from '@/components/Common/loader/loader';
+import { mainStyle } from '@/components/Common/main/styles.css';
 import { ExtendedTransactionsTable } from '@/components/extended-transactions-table/extended-transactions-table';
-import Loader from '@/components/loader/loader';
-import { mainStyle } from '@/components/main/styles.css';
-import { Text } from '@components/text';
-import Head from 'next/head';
 import React from 'react';
 
 const Transactions: React.FC = () => {
@@ -17,17 +15,7 @@ const Transactions: React.FC = () => {
 
   return (
     <div style={{ padding: '0 50px 30px 50px' }}>
-      <Head>
-        <title>Kadena Graph Client</title>
-        <link rel="icon" href="/favicon.png" />
-      </Head>
       <main className={mainStyle}>
-        <Text
-          as="h1"
-          css={{ display: 'block', color: '$mauve12', fontSize: 48, my: '$12' }}
-        >
-          Kadena Graph Client
-        </Text>
         <div>
           {loading && (
             <div style={{ display: 'flex', alignItems: 'center' }}>
